third_party/WebKit/LayoutTests/fast/canvas/canvas-getImageData-invalid-expected.txt

Test the handling of invalid arguments in canvas getImageData().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.getImageData(NaN, 10, 10, 10) threw exception TypeError: Failed to execute 'getImageData' on 'CanvasRenderingContext2D': The provided float value is non-finite..
PASS ctx.getImageData(10, NaN, 10, 10) threw exception TypeError: Failed to execute 'getImageData' on 'CanvasRenderingContext2D': The provided float value is non-finite..
PASS ctx.getImageData(10, 10, NaN, 10) threw exception TypeError: Failed to execute 'getImageData' on 'CanvasRenderingContext2D': The provided float value is non-finite..
PASS ctx.getImageData(10, 10, 10, NaN) threw exception TypeError: Failed to execute 'getImageData' on 'CanvasRenderingContext2D': The provided float value is non-finite..
PASS ctx.getImageData(Infinity, 10, 10, 10) threw exception TypeError: Failed to execute 'getImageData' on 'CanvasRenderingContext2D': The provided float value is non-finite..
PASS ctx.getImageData(10, Infinity, 10, 10) threw exception TypeError: Failed to execute 'getImageData' on 'CanvasRenderingContext2D': The provided float value is non-finite..
PASS ctx.getImageData(10, 10, Infinity, 10) threw exception TypeError: Failed to execute 'getImageData' on 'CanvasRenderingContext2D': The provided float value is non-finite..
PASS ctx.getImageData(10, 10, 10, Infinity) threw exception TypeError: Failed to execute 'getImageData' on 'CanvasRenderingContext2D': The provided float value is non-finite..
PASS ctx.getImageData(undefined, 10, 10, 10) threw exception TypeError: Failed to execute 'getImageData' on 'CanvasRenderingContext2D': The provided float value is non-finite..
PASS ctx.getImageData(10, undefined, 10, 10) threw exception TypeError: Failed to execute 'getImageData' on 'CanvasRenderingContext2D': The provided float value is non-finite..
PASS ctx.getImageData(10, 10, undefined, 10) threw exception TypeError: Failed to execute 'getImageData' on 'CanvasRenderingContext2D': The provided float value is non-finite..
PASS ctx.getImageData(10, 10, 10, undefined) threw exception TypeError: Failed to execute 'getImageData' on 'CanvasRenderingContext2D': The provided float value is non-finite..
PASS ctx.getImageData(10, 10, 0, 10) threw exception IndexSizeError: Failed to execute 'getImageData' on 'CanvasRenderingContext2D': The source width is 0..
PASS ctx.getImageData(10, 10, 10, 0) threw exception IndexSizeError: Failed to execute 'getImageData' on 'CanvasRenderingContext2D': The source height is 0..
PASS successfullyParsed is true

TEST COMPLETE
